requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
crispy-bootstrap4==2025.6
Django==5.2
django-allauth==65.13.1
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-crispy-forms==2.5
django-js-asset==3.1.2
django-parler==2.3
django-rosetta==0.10.2
idna==3.11
pillow==12.0.0
polib==1.2.0
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
